# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "API do sistema de livros com autores e assuntos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/assuntos": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Lista todos os assuntos",
                "operationId": "d508d3604432e32082c9f627ffd78c66",
                "responses": {
                    "200": {
                        "description": "Lista de assuntos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Assunto"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no banco de dados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Cria um novo assunto",
                "operationId": "af3c6b92c90b765ccc4e3cd5dfb28ecd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assunto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar assunto"
                    }
                }
            }
        },
        "/api/assuntos/{id}": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Exibe um assunto específico",
                "operationId": "4c0ede525e9d464f2ce8abeaf7c25881",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do assunto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    },
                    "500": {
                        "description": "Erro ao buscar assunto"
                    }
                }
            },
            "put": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Atualiza um assunto",
                "operationId": "dce96c0047f080f7699e13d213c666ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assunto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    },
                    "500": {
                        "description": "Erro ao atualizar assunto"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Deleta um assunto",
                "operationId": "89bb3841a5762d0a790496a61be2cecd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assunto deletado com sucesso"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    },
                    "500": {
                        "description": "Erro ao deletar assunto"
                    }
                }
            }
        },
        "/api/assuntos/{id}/restore": {
            "post": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Restaura um assunto deletado",
                "operationId": "19f843ba4345ec4de87b21dce8e90180",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assunto restaurado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao restaurar assunto"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Assunto": {
                "title": "Assunto",
                "required": [
                    "id",
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Filosofia"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T12:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}